Add separate project for tests
