--- csharpcore/GuildedRose/GildedRoseTest.cs
@@ -1,17 +1,0 @@
-﻿using Xunit;
-using System.Collections.Generic;
-
-namespace GuildedRose
-{
-    public class GildedRoseTest
-    {
-        [Fact]
-        public void foo()
-        {
-            IList<Item> Items = new List<Item> { new Item { Name = "foo", SellIn = 0, Quality = 0 } };
-            GildedRose app = new GildedRose(Items);
-            app.UpdateQuality();
-            Assert.Equal("fixme", Items[0].Name);
-        }
-    }
-}
