+++ csharpcore/GuildedRoseTests/ApprovalTest.cs
@@ -4,8 +4,9 @@ using System.IO;
 using System.Text;
 using ApprovalTests;
 using ApprovalTests.Reporters;
+using GuildedRose;
 
-namespace GuildedRose
+namespace GuildedRoseTests
 {
     [UseReporter(typeof(DiffReporter))]
     public class ApprovalTest
