+++ csharpcore/GuildedRoseTests/GildedRoseTest.cs
@@ -1,0 +1,18 @@
+﻿using Xunit;
+using System.Collections.Generic;
+using GuildedRose;
+
+namespace GuildedRoseTests
+{
+    public class GildedRoseTest
+    {
+        [Fact]
+        public void foo()
+        {
+            IList<Item> Items = new List<Item> { new Item { Name = "foo", SellIn = 0, Quality = 0 } };
+            GildedRose app = new GildedRose(Items);
+            app.UpdateQuality();
+            Assert.Equal("fixme", Items[0].Name);
+        }
+    }
+}
